split FeatureSpecHelper from IntegrationSpecHelper
the methods in the former can't be used in request specs
