+++ spec/support/integration_spec_helper.rb
@@ -19,14 +19,4 @@ module IntegrationSpecHelper
       }
     )
   end
-
-  def login_with_oauth
-    user = FactoryGirl.create(:user)
-    login_as(user)
-  end
-
-  def login_as(user)
-    setup_mock_auth(:myusa, user)
-    visit '/auth/myusa'
-  end
 end
